Implement password reset functionality
Added functions for allowing users to reset their passwords via email. The code introduces new scenes for resetting passwords and getting new passwords. Adjustments to user and product models were also made, and dayjs was added for handling date/time functionality. A password reset function was also added, which includes security checks, such as checking for token expiration. A link is printed in the console to copy and paste, which will bypass email sending for now. Changes to the routing to enable the new password reset process were also included.